Extract the Video component into its own file
